+++ src/screens/nc_help.h
@@ -25,27 +25,27 @@ struct NcHelpScreen {
     NcmBuffer search_constraint;
     NcHelpHooks hooks;
 
-    int64 lines_scrolled;
+    int32 lines_scrolled;
 };
 
 void nc_help_screen_init(NcHelpScreen *screen,
                          NcHelpHooks hooks,
-                         int64 start_x, int64 width,
-                         int64 main_start_y, int64 main_height,
+                         int32 start_x, int32 width,
+                         int32 main_start_y, int32 main_height,
                          NcColor color, NcBorder border,
-                         int64 lines_scrolled);
+                         int32 lines_scrolled);
 void nc_help_screen_set_geometry(NcHelpScreen *screen,
-                                 int64 start_x, int64 width,
-                                 int64 main_start_y,
-                                 int64 main_height);
+                                 int32 start_x, int32 width,
+                                 int32 main_start_y,
+                                 int32 main_height);
 bool nc_help_screen_reload(NcHelpScreen *screen);
 bool nc_help_screen_find(NcHelpScreen *screen, char *pattern,
                          int32 pattern_len, NcmError *error);
 void nc_help_screen_clear_search(NcHelpScreen *screen);
 NcScreen *nc_help_screen_base(NcHelpScreen *screen);
-int64 nc_help_screen_start_x(NcHelpScreen *screen);
-int64 nc_help_screen_start_y(NcHelpScreen *screen);
-int64 nc_help_screen_width(NcHelpScreen *screen);
-int64 nc_help_screen_height(NcHelpScreen *screen);
+int32 nc_help_screen_start_x(NcHelpScreen *screen);
+int32 nc_help_screen_start_y(NcHelpScreen *screen);
+int32 nc_help_screen_width(NcHelpScreen *screen);
+int32 nc_help_screen_height(NcHelpScreen *screen);
 
 #endif /* NCMPCPP_NC_HELP_H */
